major cleanup, general refactor
general fixes and cleanup, implement enums where necessary. break out styles into separate files, clean up directory structure, delete some components
