+++ src/components/Auth/components/Login/index.tsx
@@ -1,19 +1,19 @@
 import { Link } from "react-router-dom"
 
-import Form, { FormFieldType } from "../../../Form/Form"
+import Form, { FormFieldType, InputTypes } from "../../../Form/Form"
 import { PageTitle } from "../../../PageTitle"
 import { AuthContainer } from "../AuthContainer"
 import { headingMarginBottom } from "../common/common.constants"
 
 const loginForm: FormFieldType[] = [
     {
-        type: "text",
+        type: InputTypes.TEXT,
         label: "Email",
         placeholder: "Enter your username...",
         name: "email"
     },
     {
-        type: "password",
+        type: InputTypes.PASSWORD,
         label: "Password",
         placeholder: "Enter your password...",
         name: "password"
