+++ src/components/Auth/components/Register/index.tsx
@@ -1,25 +1,25 @@
 import { Link } from "react-router-dom"
 
-import Form, { FormFieldType } from "../../../Form/Form"
+import Form, { FormFieldType, InputTypes } from "../../../Form/Form"
 import { PageTitle } from "../../../PageTitle"
 import { AuthContainer } from "../AuthContainer"
 import { headingMarginBottom } from "../common/common.constants"
 
 const registerForm: FormFieldType[] = [
     {
-        type: "text",
+        type: InputTypes.TEXT,
         label: "Email Address",
         placeholder: "Enter your username...",
         name: "email"
     },
     {
-        type: "password",
+        type: InputTypes.PASSWORD,
         label: "Password",
         placeholder: "Enter your password...",
         name: "password"
     },
     {
-        type: "password",
+        type: InputTypes.PASSWORD,
         label: "Confirm Password",
         placeholder: "Confirm your password...",
         name: "password"
