+++ src/components/Layout/Footer/index.tsx
@@ -1,13 +1,14 @@
-import styled from "styled-components"
 import { Link } from "react-router-dom"
 
-import Logo from "../../assets/images/logo-final.png"
-import SocialLinks from "../SocialLinks"
-import { Col, Container, Row } from "./Grid"
+import Logo from "../../../assets/images/logo-final.png"
+import SocialLinks from "../../SocialLinks"
+import { Col, Container, Row } from "../Grid"
 
-const Footer = () => {
+import * as S from "./styles"
+
+export const Footer = () => {
     return (
-        <StyledFooter>
+        <S.Footer>
             <Container>
                 <Row>
                     <Col className="f-logos" lg={4} md={4} sm={12}>
@@ -73,66 +74,6 @@ const Footer = () => {
                     rights reserved.
                 </p>
             </div>
-        </StyledFooter>
+        </S.Footer>
     )
 }
-
-const StyledFooter = styled.div`
-    display: flex;
-    flex-flow: column;
-    border-top: 1px solid #eee;
-    margin-top: auto;
-
-    .container {
-        padding: ${({ theme }) => theme.padding.gap} 0;
-
-        .col {
-            & > img {
-                margin-bottom: 42px;
-                max-width: 90%;
-            }
-
-            &.f-logos {
-                p {
-                    margin-bottom: 12px;
-                }
-            }
-
-            &.footer-nav {
-                & > div {
-                    display: flex;
-                    flex-flow: column nowrap;
-                    margin-left: auto;
-
-                    p {
-                        margin-bottom: 18px;
-                        font-weight: bold;
-                    }
-
-                    ul {
-                        display: flex;
-                        flex-flow: column nowrap;
-                        gap: 18px;
-                        list-style: none;
-                        padding: 0;
-                        margin-left: auto;
-                    }
-                }
-            }
-        }
-    }
-
-    .copyright {
-        border-top: 1px solid ${({ theme }) => theme.colors.gray};
-        display: flex;
-        padding: 24px 0;
-
-        p {
-            color: #abadbc;
-            margin: auto;
-            font-size: ${({ theme }) => theme.fontSize.xs};
-        }
-    }
-`
-
-export default Footer
