+++ src/components/Layout/Header/components/Navigation.tsx
@@ -1,8 +1,8 @@
 import styled from "styled-components"
 import { Link } from "react-router-dom"
 
-import { Container, Col, Row } from "../Grid"
-import NavData from "../../../static/navigation.json"
+import { Container, Col, Row } from "../../Grid"
+import NavData from "../../../../static/navigation.json"
 
 type DropdownLinkType = {
     to: string
