+++ src/components/Layout/Header/components/ProfileOptions.tsx
@@ -1,7 +1,7 @@
 import styled from "styled-components"
 import { Link } from "react-router-dom"
 
-import ProfileIcon from "../../../assets/images/ico-profile.svg"
+import ProfileIcon from "../../../../assets/images/ico-profile.svg"
 
 const ProfileOptions = () => {
     return (
