+++ src/components/Layout/styles.ts
@@ -1,0 +1,12 @@
+import styled from "styled-components"
+
+export const Layout = styled.div`
+    display: flex;
+    flex-flow: column;
+    position: relative;
+    flex-grow: 2;
+
+    .placeholder {
+        min-height: 1000px;
+    }
+`
