+++ src/components/Pagination/constants.ts
@@ -1,0 +1,16 @@
+export const PAGE_SIZES = [
+    {
+        value: 6,
+        label: "6"
+    },
+    {
+        value: 12,
+        label: "12"
+    },
+    {
+        value: 24,
+        label: "24"
+    }
+]
+export const DEFAULT_PAGE_SIZE = 12
+export const PAGE_SIZE_NAME = "pagesize"
