+++ src/components/Pagination/index.tsx
@@ -1,33 +1,17 @@
 import { useEffect, useState } from "react"
-import styled from "styled-components"
 
 import { PaginationType } from "../../shared/types"
-import Select from "../Form/Select"
 import {
     ProductsFilterDataType,
     ProductsFilterDataValueType
 } from "../store/common/types.common"
-import { Pager } from "./Pager"
+import { DEFAULT_PAGE_SIZE, PAGE_SIZES, PAGE_SIZE_NAME } from "./constants"
+import { Pager } from "./components/Pager"
+import Select from "../Form/Select"
 
-const pageSizes = [
-    {
-        value: 6,
-        label: "6"
-    },
-    {
-        value: 12,
-        label: "12"
-    },
-    {
-        value: 24,
-        label: "24"
-    }
-]
+import * as S from "./styles"
 
-const DEFAULT_PAGE_SIZE = 12
-const PAGE_SIZE_NAME = "pagesize"
-
-const Pagination = ({
+export const Pagination = ({
     total,
     onChange,
     pageTotal
@@ -59,14 +43,14 @@ const Pagination = ({
     }
 
     return (
-        <StyledPagination>
+        <S.Pagination>
             <Pager settings={page} total={total} onChange={handlePaginate} />
             <p className="fade">{`Items ${pageTotal} of ${total}`}</p>
             <div className="show">
                 <p>Show</p>
                 <Select
                     name="pagesize"
-                    options={pageSizes}
+                    options={PAGE_SIZES}
                     defaultValue={{
                         name: PAGE_SIZE_NAME,
                         value: DEFAULT_PAGE_SIZE
@@ -76,32 +60,6 @@ const Pagination = ({
                     }
                 />
             </div>
-        </StyledPagination>
+        </S.Pagination>
     )
 }
-
-const StyledPagination = styled.div`
-    display: flex;
-    flex-flow: row nowrap;
-    gap: 32px;
-    border: 1px solid ${({ theme }) => theme.colors.gray};
-    padding: 12px 24px;
-    border-radius: ${({ theme }) => theme.borderRadius};
-
-    .fade {
-        color: ${({ theme }) => theme.colors.text.faded};
-    }
-
-    p {
-        margin: auto 0;
-        font-size: ${({ theme }) => theme.fontSize.sm};
-    }
-
-    .show {
-        display: flex;
-        flex-flow: row nowrap;
-        gap: 12px;
-    }
-`
-
-export default Pagination
